** Fixed issue #242 and added tests to back it up

+++ webapp/src/test/resources/rest-user-client-teardown.sql
@@ -1,3 +1,3 @@
-delete from OAuthAccessToken where accessToken = 'access-token';
-delete from OAuthAuthorizationCode where client_id = (select id from OAuthClient where name = 'test-user-client');
-delete from OAuthClient where name = 'test-user-client';
+delete from OAuthAccessToken where accessToken in ('access-token', 'admin-access-token');
+delete from OAuthAuthorizationCode where client_id in (select id from OAuthClient where name in ('test-user-client', 'test-admin-client'));
+delete from OAuthClient where name in ('test-user-client', 'test-admin-client');
